+++ src/main/java/com/gruppe21/card/cardControllers/UseOnDrawCardController.java
@@ -1,7 +1,7 @@
 package com.gruppe21.card.cardControllers;
 
 import com.gruppe21.card.cardType.Card;
-import com.gruppe21.card.cardView.CardView;
+import com.gruppe21.card.view.CardView;
 import com.gruppe21.player.PlayerController;
 
 public abstract class UseOnDrawCardController extends CardController{
