+++ src/main/java/com/gruppe21/card/moveCards/controllers/MoveRelativeCardController.java
@@ -1,8 +1,8 @@
-package com.gruppe21.card.cardControllers;
+package com.gruppe21.card.moveCards.controllers;
 
 
-import com.gruppe21.card.cardType.MoveRelativeCard;
-import com.gruppe21.card.cardView.CardView;
+import com.gruppe21.card.moveCards.model.MoveRelativeCard;
+import com.gruppe21.card.view.CardView;
 import com.gruppe21.game.Board;
 import com.gruppe21.player.PlayerController;
 import com.gruppe21.squares.controllers.SquareController;
