+++ src/main/java/com/gruppe21/card/moveCards/model/TeleportToNearestCard.java
@@ -1,5 +1,6 @@
-package com.gruppe21.card.cardType;
+package com.gruppe21.card.moveCards.model;
 
+import com.gruppe21.card.cardType.Card;
 import com.gruppe21.squares.controllers.SquareController;
 import org.w3c.dom.Element;
